useConvertToGroupButtonProps hook
