+++ packages/block-editor/src/components/block-settings-menu-controls/index.js
@@ -12,7 +12,11 @@ import { useSelect } from '@wordpress/data';
 /**
  * Internal dependencies
  */
-import ConvertToGroupButton from '../convert-to-group-buttons';
+import {
+	useConvertToGroupButtonProps,
+	ConvertToGroupButton,
+} from '../convert-to-group-buttons';
+import { store as blockEditorStore } from '../../store';
 
 const { Fill: BlockSettingsMenuControls, Slot } = createSlotFill(
 	'BlockSettingsMenuControls'
@@ -22,7 +26,7 @@ const BlockSettingsMenuControlsSlot = ( { fillProps, clientIds = null } ) => {
 	const selectedBlocks = useSelect(
 		( select ) => {
 			const { getBlocksByClientId, getSelectedBlockClientIds } = select(
-				'core/block-editor'
+				blockEditorStore
 			);
 			const ids =
 				clientIds !== null ? clientIds : getSelectedBlockClientIds();
@@ -34,14 +38,20 @@ const BlockSettingsMenuControlsSlot = ( { fillProps, clientIds = null } ) => {
 		[ clientIds ]
 	);
 
+	// Check if current selection of blocks is Groupable or Ungroupable
+	// and pass this props down to ConvertToGroupButton.
+	const convertToGroupButtonProps = useConvertToGroupButtonProps();
+	const { isGroupable, isUngroupable } = convertToGroupButtonProps;
+	const showConvertToGroupButton = isGroupable || isUngroupable;
 	return (
 		<Slot fillProps={ { ...fillProps, selectedBlocks } }>
 			{ ( fills ) => {
-				if ( fills?.length > 0 ) {
+				if ( fills?.length > 0 || showConvertToGroupButton ) {
 					return (
 						<MenuGroup>
 							{ fills }
 							<ConvertToGroupButton
+								{ ...convertToGroupButtonProps }
 								onClose={ fillProps?.onClose }
 							/>
 						</MenuGroup>
